feat: video cli tool
Adds a video cli tool which allows for configuration of the video api
from a simple to use CLI interface.

+++ video/common/src/database/room.rs
@@ -55,7 +55,7 @@ pub struct Room {
 	pub active_recording_id: Option<Ulid>,
 
 	/// The ingest bitrate
-	pub ingest_bitrate: Option<i32>,
+	pub ingest_bitrate: Option<i64>,
 
 	/// The video output configs after transcoding
 	pub video_output: Option<Vec<Protobuf<VideoConfig>>>,
